backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==1.4.41
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0 